Update README.md
I found it difficult to figure out how to pass an empty third argument, I had to change back to `master` branch for how I am using this action. I applaud the idea of defaulting to `main` and this gives me further reason to make that transition.  I was managing about quite a few different starters via this.  Just included an example here.